Changed git ignore file and found parsing error in vimrc
Added .dSYS to the list of files that git should ignore when it asks
about things to add to the repository. My other change was the discovery
that I had somehow made a character in my vimrc that was causing vim to be
unable to load my specific c code formatting instructions for .c files.
It was throwing an "E20: Mark not set" error. Retyping the line fixed the
problem. Mysteries of life...