Removed both Supertab and cppomnicomplete.
So, I'm going to be moving to using the YoucompleteMe plugin.
As part of this process I'm going to be removing my Supertab and
other .cpp omnicompletion settings. Hopefully, this ends up being
a useful choice. We shall see... Plus side of using a version control
system is that I can also roll back right?